+++ src/Core/Categories/Events/CategoryStoredEvent.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace GetCandy\Api\Core\Categories\Events;
+
+use Illuminate\Broadcasting\Channel;
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Broadcasting\PrivateChannel;
+use GetCandy\Api\Core\Categories\Models\Category;
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Broadcasting\InteractsWithSockets;
+
+class CategoryStoredEvent
+{
+    use Dispatchable, InteractsWithSockets, SerializesModels;
+
+    protected $category;
+
+    /**
+     * Create a new event instance.
+     *
+     * @return void
+     */
+    public function __construct(Category $category)
+    {
+        $this->category = $category;
+    }
+
+    public function category()
+    {
+        return $this->category;
+    }
+
+    /**
+     * Get the channels the event should broadcast on.
+     *
+     * @return Channel|array
+     */
+    public function broadcastOn()
+    {
+        return new PrivateChannel('products');
+    }
+}