+++ src/Core/Payments/Events/PaymentFailedEvent.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace GetCandy\Api\Core\Payments\Events;
+
+use Illuminate\Broadcasting\Channel;
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Broadcasting\PrivateChannel;
+use GetCandy\Api\Core\Baskets\Models\Basket;
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Broadcasting\InteractsWithSockets;
+
+class PaymentFailedEvent
+{
+    use Dispatchable, InteractsWithSockets, SerializesModels;
+
+    public $errors;
+
+    /**
+     * Create a new event instance.
+     *
+     * @return void
+     */
+    public function __construct($errors)
+    {
+        $this->errors = $errors;
+    }
+
+    /**
+     * Get the channels the event should broadcast on.
+     *
+     * @return Channel|array
+     */
+    public function broadcastOn()
+    {
+        return new PrivateChannel('payments');
+    }
+}